pyproject.toml
[tool.poetry]
name = "poly-bench"
version = "0.1.0"
description = "Polysemantic benchmark transformers"
authors = ["Evan Anders <evanhanders@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
datasets = "^2.21.0"
einops = "^0.8.0"
jaxtyping = "^0.2.33"
safetensors = "^0.4.4"
transformers = "^4.44.1"
numpy = "<2.0.0"
huggingface-hub = ">=0.23.0"
jupyter = "^1.0.0"
iit = { git = "https://github.com/evanhanders/iit.git", branch = "training_improvements"}
pyyaml = "^6.0.2"
mypy = "^1.11.1"
types-pyyaml = "^6.0.12.20240808"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
